# api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: VionetaOS App Management API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/Vioneta/logo/main/vionetaos/vionetaos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/Vioneta/logo/main/vionetaos/vionetaos_banner_twilight_blue_800px.png">
        <img alt="VionetaOS" src="https://raw.githubusercontent.com/Vioneta/logo/main/vionetaos/vionetaos_banner_twilight_blue_800px.png">
    </picture>

    VionetaOS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.

    For issues and discussions, please visit the [GitHub repository](https://github.com/Vioneta/VionetaOS) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
- url: /v2/app_management
...
security:
- access_token: []
paths:
...
  /global:
    get:
      summary: Get all global settings
      operationId: getGlobalSettings
      tags:
      - Common methods
      responses:
        "200":
          $ref: "#/components/responses/GlobalSettingsOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /global/{key}:
    put:
      summary: Update a global setting
      operationId: updateGlobalSetting
      tags:
      - Common methods
      parameters:
      - $ref: "#/components/parameters/GlobalSettingKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalSetting"
      responses:
        "200":
          $ref: "#/components/responses/GlobalSettingOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    delete:
      summary: Delete a global setting
      operationId: deleteGlobalSetting
      tags:
      - Common methods
      parameters:
      - $ref: "#/components/parameters/GlobalSettingKey"
      responses:
        "200":
          $ref: "#/components/responses/ResponseOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    get:
      summary: Get a global setting
      operationId: getGlobalSetting
      tags:
      - Common methods
      parameters:
      - $ref: "#/components/parameters/GlobalSettingKey"
      responses:
        "200":
          $ref: "#/components/responses/GlobalSettingOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    GlobalSettingKey:
      name: key
      description: Key of the global setting
      in: path
      required: true
      schema:
        type: string
        example: "OPENAI_API_KEY"
...
  responses:
    ResponseOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "OK"
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
    ResponseBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Bad Request"
...
    GlobalSettingOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
            - $ref: "#/components/schemas/BaseResponse"
            - properties:
                data:
                  $ref: "#/components/schemas/GlobalSetting"
    GlobalSettingsOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
            - $ref: "#/components/schemas/BaseResponse"
            - properties:
                data:
                  type: array
                  items:
                    $ref: "#/components/schemas/GlobalSetting"
...
  schemas:
    BaseResponse:
      properties:
        message:
          description: message returned by server side if there is any
          type: string
          example: ""
...
    GlobalSetting:
      required:
      - key
      - value
      properties:
        key:
          description: key of the setting
          type: string
          readOnly: true
          example: "OPENAI_API_KEY"
        value:
          description: value of the setting
          type: string
          example: "sk-xxxxxxx"
        description:
          description: description of the setting
          type: string
          example: "OpenAI API Key"
...